Add ApiClient with `track` function
This adds a new class, an ApiClient, that has the same interface as the
client in Flutter (minus the batch tracker, as that will come later).

It takes in a single event and posts it to the server. The track
function returns a boolean on whether it was successful or not.