Add Appfit Cache
This adds an Appfit Cache class that interacts with local storage.

It stores a user id to a local storage cache, if provided, or to an in-memory cache
if for some reason local storage is unavailable. In the case that a
user id is not included, it instead installs a random uuid that serves
as a temporary anonymous user id.