Add EventCache for failed events
This adds a reduced version of the EventCache present in the mobile
SDKs, which save dropped events to disk.

Since this is aimed at the web, we won't be persisting these to disk.
However, we can at least keep them in memory. These can be batch
processed when internet is restored.