Add event digester
This adds the event digester, which has the main logic of taking events
from the `track` function and consuming them appropriately. It converts
them to dtos and sends them to api client.

It also caches any failed events and retries them with the batch
endpoint any time there's a successful digestion of a single event.

I also made a few adjustments to the appfit cache to be clearer and
provide all the necessary logic.