Add event digester integration tests
This adds a new test file that is an integration test for the event
digester. Since this is the entry point for all logic (the AppFit class
is just a public interface), it is an integration test for all
functionality.

The only thing it mocks are two external dependencies:
http calls and uuid generation. However, it also sets the local storage
availability flag to false, since that doesn't exist in a test
environment.