Small build process improvements
This adds a few small improvements to the build process. It adds to the
package.json an "export" property to be more current. It also adds a
"files" property for white-listing our build directory for publishing.
Finally, I add a prepublish script that is run before a publish.

I also removed the stub code from the index and replaced it with
re-exporting the public interface