Move browser code into package folder for monorepo
I am changing up this project to include two packages: a frontend and
a backend sdk. We have decided to use a monorepo to hold the code for
both. This starts re-organizing in order to accomplish that.

I have moved the code that was originally in the `src` folder into a
new `packages` folder instead. I have duplicated the build configs so
that the individual package has its own. This allows us to use npm
workspaces to create a build.