Refactor core appfit logic into shared folder
This pulls all the core appfit business logic out of the browser sdk
and moves it instead into the shared folder. The purpose is that this
code can then be shared in both the browser and the node sdks.

In the browser sdk package I have left the things that are specific to
browser clients: the caching of the user and events and the interface
that wraps the internal event digester.

Everything else is moved to the shared package.