Add Parcel as bundler
This adds the new root package Parcel in order to facilitate bundling
during builds. This is a zero-config bundler, which is probably fine
for an sdk of our size (we might need to switch later if we need
something that is more configurable).

To get it working, I also needed to update the package jsons to add a
source property and to update the build scripts. I also had to change
some of the re-exports in the index files to specifically mark them as
`type` exports.