Add browser wrapper script for loading AppFit from `script` tag
We need a different public interface for loading AppFit when it is done from inside `script` tags in a browser rather than in a web app.

To do this I have added a new `snippit.html` file that contains a script to create a temporary AppFit global object that loads the main script from a CDN. It caches events and identify calls until the script is actually loaded so that we don't miss anything.

I also changed the shared code to be imported directly rather than as a package, so that it could all be bundled for the browser.

I also made a few fixes that I caught now that I am actually able to test the loading.