Add builds for legacy browsers
This makes a few changes in order to correctly build for legacy browsers
in addition to modern browsers.

I updated the config for Parcel so that it now makes two browser bundles,
one for modern and one for legacy.

I removed the multiple tsconfig files that were used by `tsc` to make
multiple outputs, since I'm now relying entirely on Parcel.

I updated the snippet and test html pages to have both the main script
and the legacy fallback.