Add buildspec files for code build
This adds buildspec YAML files for specifying to AWS Codebuild how to
build and publish the SDK. It also makes some updates to the package
json files to improve the install scripts.

The dev and prod buildspecs are the same except that they upload the
browser version of the sdk to different buckets, and only the prod build
publishes packages to npm.

I updated the lerna json file to allow the packages to be versioned
independently. I also updated the scripts slightly

<!-- ps-id: ba4ddb46-22db-48b1-8e2f-09e3c81d1325 -->