Reorganize files in repo
I realized while trying to fix an error with the browser imports that it
would be better to separate the cdn build of the sdk from the browser
sdk. To that end, I made it into its own separate module.

At the same time, I took the liberty to rename the folders, to simplify
them and remove duplication in the names.

I also removed relative imports and instead used the packages. This
still works for the cdn build, as it automatically bundles the
dependencies in. (I'm still not clear if this will work with importing
the browser packages without actually publishing the shared library to
npm as well.)