Add origin to system properties
This is a somewhat transitional change that adds an option to pass in
an 'origin' to the CDN version of appfit and that updates the snippet
to pass in `'web'` as that origin. It is transitional because it
is very likely not the final version of interface for the CDN, and so
that means we will have another major update soon.

To accomplish this I had to make a number of changes. First, the CDN
interface was a wrapper before, but because this needs to be passed in,
I had to make a separate version of the AppFit class to act as an
interface. It is exactly the same except that it adds the option to pass
in the origin string. I made a new interface because in the future we
plan to add plugins, which will dictate a different interface anyway.
Also, since the origin is optional, old versions of the snippet will
still work fine.

In order to accommodate both versions of the AppFit public class, I
moved some of the other classes into the shared package. I also moved
the system events deeper into the infrastructure.

Finally, I switched the CDN index to have the `appFitInstall` function
be an IIFE so it doesn't have to stay in the global scope.