Update CDN destination file names
This updates the file names of the non-versioned CDN file to
include a `v1` tag. The reason for this is that we decided that this
file can be updated as long as it sticks to the convention of not
having major version changes. (We can make a new v2 file once we change
something major.) This reinforces that by actually adding the major
tag to the file name.

In the future, only the buildspec file needs to be changed to v2 once
we have that changed.

I also updated the dev build to match the directory location that we
have in prod.