Refactor appfit-shared to prepare for server package
This makes several changes in preparation for the node/server package.
This first thing it does it make a new AppFitCore class that serves as
the main entry point for the shared library instead of the event
digester.

The event digester was doing too much, so some of its code I moved into
AppFitCore. It is now only in charge of sending events to the API and
caching them/retrying in the case of failure. It doesn't "know" about
users anymore.

AppFitCore optionally accepts a user cache, which it will use if user
information for a particular event is not explicitly included. The
server will not be using a cache.

I also made a new event type, AnalyticEvent, which is now the main
internal business model for events (leaving AppFitEvent to be a public
interface). It holds user and time information, so that if something
fails, events will not be misattributed to a user if they were
re-identified.

Finally, I refactored the browser & cdn wrappers to switch to
AppFitCore and updated the tests.

<!-- ps-id: 77871a6a-b3f5-4c49-b6d2-ac6fda8190d8 -->