Add browser user agent to browser/cdn sdks
Since we want to ingest the user agent from the user's browser when they
are using one, this automatically fetches the user agent if available
and adds it to the system properties of the browser and CDN SDKs. It
does this by allowing system properties to be passed into AppFit Core.

<!-- ps-id: 90104aca-8465-42da-a6f2-885856ea8055 -->