Add config options with app version property
This adds a new, optional configuration options parameter to the
config class of all three SDKs. This will allow us to expose options
to the user for various configurations.

For now, I am including only a single config option, which is the app
version. This allows the user to set what version of their own app
they are using. This gets included in the 'systemProperties' part of
an event

[changelog]
added: configuration options with app version property

<!-- ps-id: 12c642af-0d7e-4c17-87e8-1b7a062c9494 -->