Add browser option to enable ip tracking
This adds a new option parameter to the browser config options that
will allow the enabling or disabling of ip address tracking. This is not
yet implemented.

I left out the option for the server and have it set to disabled
automatically, since there's no point.

Finally, i bumped the node version.

[changelog]
added: option to enable or disable ip tracking in browser

<!-- ps-id: 4b17971b-701d-41f8-bca6-8f5ece5dd4ba -->