Cache ip address with user cache
This updates the browser user cache to also be able to persist the
ip address. It does this by exposing two new functions. The ip address
caching expires after a specified time.

I did accomplished this by also saving an expiration time when the ip
address is cached, and then checking that time when the value is read.
If the value is expired, we delete the cache and return `undefined` as
if it was not saved.