Fetch IP address in appfit core if not disabled
This change has appfit core read the option config for enabling ip
tracking, and if it is not disabled, we fetch the ip address from a
different site and add it to tracking events. The ip address is cached
for one hour before expiring.

To do this I made a new ip address client which simply fetches an ip
address from a third party. This client is optionally injected into
appfit core, so both browser wrappers inject it while the server does
not.

Appfit core tracking has been updated to grab the ip address from the
user cache or re-fetch it if it is missing.